Fix setStyleState regression in 874ec79
Oh React. If you call `setState` twice during the same execution frame,
the second time, `this.state` will be the same as the first time. Thus,
the second time will clobber the changes made the first time. This is a
known issue: https://github.com/facebook/react/issues/122

Btw, I noticed this when running the examples and trying the one with
all the boxes; mouseLeave on one box gets called in the same frame as
mouseEnter on another, and since all the boxes are in one component,
they clobbered each others state, and everyone stayed hovered. Too lazy
to write a proper test for this right now.