Use Object.getOwnPropertyNames instead of Object.keys
From @Zenwolf's comment:

ES6 class properties are not enumerable with Object.keys, as per the spec. They are defined with Object.defineProperty where enumerable is false by default. So the keys will not show up using Object.keys() and will not be copied. When you assign a vanilla property with MyComponent.foo = someVal, it will show up because the old default for enumerable when not using Object.defineProperty is true.